Get the extension to compile, at least. Won't load into PG yet.
This liberally copies the setup from the xml2 extension in PG 9.1's contrib/,
and the to_html() function copies liberally from RDiscount (Discount, ie
Markdown, for Ruby). The code compiles, so long as Discount is compiled and
installed to the system, and the directory is copied to the contrib/ directory
of a compiled postgresql directory. It depends on version 9.1 or greater of
PostgreSQL.

Sadly, it doesn't actually load when CREATE EXTENSION is called. Not sure why.

That should probably get fixed at some point.